+++ examples/GeneratorFormats.cpp
@@ -65,10 +65,10 @@ int main()
     std::cout << gen.xml(40) << '\n';
 
     // Generate a JSON object with a max recursion of 5.
-    std::cout << gen.json(0, 5, lpsm::OBJECT) << '\n';
+    std::cout << gen.json(5, lpsm::OBJECT) << '\n';
 
     // Generate a JSON value with a max recursion of 5.
-    std::cout << gen.json_value(0, 5) << '\n';
+    std::cout << gen.json_value(5) << '\n';
 
     return 0;
 }